agents/requirements.txt
langchain==0.1.20
langchain-google-genai==1.0.0
langgraph==0.0.55
google-generativeai==0.4.0
httpx==0.26.0
python-dotenv==1.0.0